Much refactoring Fixed Move semantics in openssl impl Use in-place construction for secp256k1 Swapped implementations in mixed - sign with libsecp256k1, verify with openssl
