+++ src/crypto/elliptic_common.cpp
@@ -1,19 +1,12 @@
+#include <fc/crypto/base58.hpp>
+#include <fc/crypto/elliptic.hpp>
+
+/* stuff common to all ecc implementations */
+
 namespace fc { namespace ecc {
-    public_key::public_key() {}
-
-    public_key::~public_key() {}
-
-    public_key::public_key( const public_key& pk ) : my( pk.my ) {}
-
-    public_key::public_key( public_key&& pk ) : my( std::move(pk.my) ) {}
 
     public_key public_key::from_key_data( const public_key_data &data ) {
         return public_key(data);
-    }
-
-    bool public_key::valid()const
-    {
-      return my->_key != nullptr;
     }
 
     std::string public_key::to_base58( const public_key_data &key )
@@ -45,14 +38,6 @@ namespace fc { namespace ecc {
                && !(c.data[33] & 0x80)
                && !(c.data[33] == 0 && !(c.data[34] & 0x80));
     }
-
-    private_key::private_key() {}
-
-    private_key::~private_key() {}
-
-    private_key::private_key( const private_key& pk ) : my(pk.my) {}
-
-    private_key::private_key( private_key&& pk ) : my( std::move( pk.my) ) {}
 
     private_key private_key::generate_from_seed( const fc::sha256& seed, const fc::sha256& offset )
     {
@@ -108,29 +93,6 @@ namespace fc { namespace ecc {
        return private_key( k );
     }
 
-    private_key& private_key::operator=( private_key&& pk )
-    {
-        my = std::move(pk.my);
-        return *this;
-    }
-
-    public_key& public_key::operator=( public_key&& pk )
-    {
-        my = std::move(pk.my);
-        return *this;
-    }
-
-    public_key& public_key::operator=( const public_key& pk )
-    {
-        my = pk.my;
-        return *this;
-    }
-
-    private_key& private_key::operator=( const private_key& pk )
-    {
-        my = pk.my;
-        return *this;
-    }
 }
 
 void to_variant( const ecc::private_key& var,  variant& vo )
@@ -156,4 +118,5 @@ void from_variant( const variant& var,  ecc::public_key& vo )
     from_variant( var, dat );
     vo = ecc::public_key(dat);
 }
+
 }
