+++ src/crypto/elliptic_mixed.cpp
@@ -10,42 +10,31 @@
 #include <assert.h>
 #include <secp256k1.h>
 
+#include "_elliptic_impl_priv.hpp"
+#include "_elliptic_impl_pub.hpp"
+
 namespace fc { namespace ecc {
     namespace detail
     {
-        static void init_lib();
+        static int init_secp256k1() {
+            secp256k1_start(SECP256K1_START_VERIFY | SECP256K1_START_SIGN);
+            return 1;
+        }
 
-        typedef public_key_data pub_data_type;
-        typedef EC_KEY priv_data_type;
-
-        #include "_elliptic_impl.cpp"
+        void _init_lib() {
+            static int init_s = init_secp256k1();
+            static int init_o = init_openssl();
+        }
     }
 
-    #include "_elliptic_mixed_openssl.cpp"
-
-    public_key private_key::get_public_key()const
-    {
-        public_key_data data;
-        EC_KEY_set_conv_form( my->_key, POINT_CONVERSION_COMPRESSED );
-        unsigned char* buffer = (unsigned char*) data.begin();
-        i2o_ECPublicKey( my->_key, &buffer ); // FIXME: questionable memory handling
-        return public_key( data );
-    }
-
+    static const private_key_secret empty_priv;
     fc::sha512 private_key::get_shared_secret( const public_key& other )const
     {
-      FC_ASSERT( my->_key != nullptr );
+      FC_ASSERT( my->_key != empty_priv );
       FC_ASSERT( other.my->_key != nullptr );
-      fc::sha512 buf;
-      EC_KEY* key = EC_KEY_new_by_curve_name( NID_secp256k1 );
-      const unsigned char* buffer = (const unsigned char*) other.my->_key->begin();
-      o2i_ECPublicKey( &key, &buffer, sizeof(*other.my->_key) );
-      ECDH_compute_key( (unsigned char*)&buf, sizeof(buf), EC_KEY_get0_public_key(key), my->_key, ecies_key_derivation );
-      EC_KEY_free(key);
-      return buf;
+      public_key_data pub(other.serialize());
+      FC_ASSERT( secp256k1_ec_pubkey_tweak_mul( (unsigned char*) pub.begin(), pub.size(), (unsigned char*) my->_key.data() ) );
+      return fc::sha512::hash( pub.begin() + 1, pub.size() - 1 );
     }
 
-    #include "_elliptic_mixed_secp256k1.cpp"
 } }
-
-#include "_elliptic_common.cpp"
